Add routes to profile page
